modernize SGlib code

+++ fe-aqua/SGTableView.h
@@ -27,7 +27,7 @@
 
 //@interface NSObject(SGTableDataSource)
 @protocol SGTableDataSource
-- (NSSize) tableView:(NSTableView *) aTableView sizeHintForTableColumn:(NSTableColumn *) aTableColumn row:(int) rowIndex;
-- (void) tableView:(NSTableView *) aTableView sizeHintForTableColumn:(NSTableColumn *) aTableColumn row:(int) rowIndex size:(NSSize) size;
+- (NSSize) tableView:(NSTableView *) aTableView sizeHintForTableColumn:(NSTableColumn *) aTableColumn row:(NSInteger) rowIndex;
+- (void) tableView:(NSTableView *) aTableView sizeHintForTableColumn:(NSTableColumn *) aTableColumn row:(NSInteger) rowIndex size:(NSSize) size;
 - (BOOL) shouldDoSizeFixupsForTableView:(NSTableView *) aTableView;
 @end
